feat: useRegisterUser hook
